Implementation of LocalizedDateTimeField
This model field accepts a timezone override which can be a timezone string, timezone object, pytz object, callable which returns any of the previous, a model queryset keyword relation (i.e. fk__timezone) or None (default). If it is None then it defaults to settings.TIME_ZONE, and defaults to 'UTC' if that is not set.

The datetime field will be stored in teh database with the settings.TIME_ZONE timezone (defaults to UTC). The backend storage should match settings.TIME_ZONE so that time based queryset relations work properly. When accessing the datetime attribute, it will be in the Timezone specified by the timezone argument (which could be a lazy callable, or a model relation.

Fixing some minor bugs with the decorator.



git-svn-id: https://django-timezones.googlecode.com/svn/trunk@22 86ebb30f-654e-0410-bc0d-7bf82786d749