Removing TimeZoneField from top-level import
This causes problems when installing the package when all the
dependencies aren't already there. setup.py is parsed - which looks for
a version string - which then runs this __init__ - which then has to
handle the imports, which if we haven't yet got our dependencies
installed then they fail.

If anyone knows a nice way to get around this, please let me know.